# go-testing/go-web-testing/docs/swagger.yaml
definitions:
...
  handler.SectionRequest:
    properties:
      current_capacity:
        type: integer
      current_temperature:
        type: integer
      maximum_capacity:
        type: integer
      minimum_capacity:
        type: integer
      minimum_temperature:
        type: integer
      product_type_id:
        type: integer
      section_number:
        type: integer
      warehouse_id:
        type: integer
    required:
    - current_capacity
    - current_temperature
    - maximum_capacity
    - minimum_capacity
    - minimum_temperature
    - product_type_id
    - section_number
    - warehouse_id
    type: object
  handler.UpdateSectionRequest:
    properties:
      current_capacity:
        type: integer
      current_temperature:
        type: integer
      maximum_capacity:
        type: integer
      minimum_capacity:
        type: integer
      minimum_temperature:
        type: integer
      product_type_id:
        type: integer
      section_number:
        type: integer
      warehouse_id:
        type: integer
    type: object
...
  handler.request:
    properties:
      address:
        type: string
      minimum_capacity:
        type: integer
      minimum_temperature:
        type: integer
      telephone:
        type: string
      warehouse_code:
        type: string
    type: object
  web.response:
    properties:
      data: {}
    type: object
info:
  contact:
    name: API Support
    url: https://localhost:8080/
  description: This API Handle MELI Products.
  termsOfService: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  title: MELI Bootcamp API FP
  version: "1.0"
paths:
  /api/v1/products:
    post:
      consumes:
      - application/json
      description: create products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to create
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/web.response'
      summary: Create products
      tags:
      - Products
  /api/v1/products/:id:
    delete:
      consumes:
      - application/json
      description: delete product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Delete product
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: update products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to update
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handler.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Update products
      tags:
      - Products
...
  /products:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: List products
      tags:
      - Products
  /products/:id:
    get:
      consumes:
      - application/json
      description: get product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.response'
      summary: Get product
      tags:
      - Products
  /sections:
    get:
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of sections
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized + invalid token
          schema:
            $ref: '#/definitions/web.response'
        "500":
          description: 'Internal server error '
          schema:
            $ref: '#/definitions/web.response'
      summary: Show the list of sections in db
      tags:
      - Sections
    post:
      consumes:
      - application/json
      parameters:
      - description: Token
        in: header
        name: token
        required: true
        type: string
      - description: Section to save
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.SectionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created section
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized + invalid token
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/web.response'
        "422":
          description: Gin validator error
          schema:
            $ref: '#/definitions/web.response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/web.response'
      summary: Save a new section in db
      tags:
      - Sections
  /sections/{id}:
    delete:
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: 'Unauthorized: empty or invalid token'
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Section not found
          schema:
            $ref: '#/definitions/web.response'
      summary: Delete a required section
      tags:
      - Sections
    get:
      consumes:
      - application/json
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Section
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: Unauthorized + invalid token
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Section not found
          schema:
            $ref: '#/definitions/web.response'
      summary: Show the required section
      tags:
      - Sections
    put:
      consumes:
      - application/json
      parameters:
      - description: Section ID
        in: path
        name: id
        required: true
        type: integer
      - description: Token
        in: header
        name: token
        type: string
      - description: Section fields to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/handler.UpdateSectionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Section
          schema:
            $ref: '#/definitions/web.response'
        "400":
          description: Bad request
          schema:
            $ref: '#/definitions/web.response'
        "401":
          description: 'Unauthorized: empty or invalid token'
          schema:
            $ref: '#/definitions/web.response'
        "404":
          description: Section no found
          schema:
            $ref: '#/definitions/web.response'
        "409":
          description: Section number is already in use
          schema:
            $ref: '#/definitions/web.response'
      summary: Update the section with the required fields
      tags:
      - Sections
swagger: "2.0"